requirements.txt
# Interfaz Web y Servidor
dash==2.18.2
dash-bootstrap-components==1.6.0
Werkzeug==3.0.6

# Procesamiento de Datos y Matemáticas
pandas==2.2.3
numpy==2.0.2
scikit-learn==1.5.2

# Machine Learning
xgboost==2.1.3
joblib==1.4.2

# IA y Agentes
google-genai==0.6.0

# Utilidades y Fechas
requests==2.32.3
holidays==0.67
python-dotenv==1.0.1

# Exportación y Reportes
openpyxl==3.1.5